Update initials in menu and fix frontend unit tests
